pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bayesian-envhealth-models"
dynamic = ["version"]
description = "Bayesian models for environmental health"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "theorashid", email = "theoaorashid@gmail.com" },
  { name = "rmp15", email = "robbie.parks@columbia.edu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Intended Audience :: Science/Research",
]
dependencies = ["hydra-core", "numpyro", "jaxtyping", "beartype", "arviz"]

[project.optional-dependencies]
dev = ["ruff", "pre-commit"]

[project.urls]
Documentation = "https://github.com/unknown/bayesian-envhealth-models#readme"
Issues = "https://github.com/unknown/bayesian-envhealth-models/issues"
Source = "https://github.com/unknown/bayesian-envhealth-models"

[tool.hatch.version]
path = "src/bayesian_envhealth_models/__init__.py"

[tool.ruff]
extend-include = ["*.ipynb"]
exclude = []
ignore = ["F821"]
